hf_deployment/requirements.txt
streamlit
tweepy
google-generativeai
requests
python-dotenv
